Hiding impl details for lap, spectral, and label
