+++ cpp/include/raft/label/classlabels.hpp
@@ -16,16 +16,10 @@
 
 #pragma once
 
-#include <cub/cub.cuh>
-
-#include <raft/cuda_utils.cuh>
-#include <raft/cudart_utils.h>
-#include <raft/linalg/unary_op.cuh>
-#include <rmm/device_scalar.hpp>
-#include <rmm/device_uvector.hpp>
+#include <raft/label/detail/classlabels.cuh>
 
 namespace raft {
-namespace label {
+    namespace label {
 
 /**
  * Get unique class labels.
@@ -42,39 +36,8 @@ namespace label {
  * \param [in] stream cuda stream
  */
 template <typename value_t>
-int getUniquelabels(rmm::device_uvector<value_t>& unique, value_t* y, size_t n, cudaStream_t stream)
-{
-  rmm::device_scalar<int> d_num_selected(stream);
-  rmm::device_uvector<value_t> workspace(n, stream);
-  size_t bytes  = 0;
-  size_t bytes2 = 0;
-
-  // Query how much temporary storage we will need for cub operations
-  // and allocate it
-  cub::DeviceRadixSort::SortKeys(
-    NULL, bytes, y, workspace.data(), n, 0, sizeof(value_t) * 8, stream);
-  cub::DeviceSelect::Unique(
-    NULL, bytes2, workspace.data(), workspace.data(), d_num_selected.data(), n, stream);
-  bytes = max(bytes, bytes2);
-  rmm::device_uvector<char> cub_storage(bytes, stream);
-
-  // Select Unique classes
-  cub::DeviceRadixSort::SortKeys(
-    cub_storage.data(), bytes, y, workspace.data(), n, 0, sizeof(value_t) * 8, stream);
-  cub::DeviceSelect::Unique(cub_storage.data(),
-                            bytes,
-                            workspace.data(),
-                            workspace.data(),
-                            d_num_selected.data(),
-                            n,
-                            stream);
-
-  int n_unique = d_num_selected.value(stream);
-  // Copy unique classes to output
-  unique.resize(n_unique, stream);
-  raft::copy(unique.data(), workspace.data(), n_unique, stream);
-
-  return n_unique;
+int getUniquelabels(rmm::device_uvector<value_t>& unique, value_t* y, size_t n, cudaStream_t stream) {
+    detail::getUniqueLabels<value_t>(unique, y, n, stream);
 }
 
 /**
@@ -97,45 +60,9 @@ int getUniquelabels(rmm::device_uvector<value_t>& unique, value_t* y, size_t n, 
  */
 template <typename value_t>
 void getOvrlabels(
-  value_t* y, int n, value_t* y_unique, int n_classes, value_t* y_out, int idx, cudaStream_t stream)
-{
-  ASSERT(idx < n_classes,
-         "Parameter idx should not be larger than the number "
-         "of classes");
-  raft::linalg::unaryOp(
-    y_out,
-    y,
-    n,
-    [idx, y_unique] __device__(value_t y) { return y == y_unique[idx] ? +1 : -1; },
-    stream);
-  RAFT_CUDA_TRY(cudaPeekAtLastError());
+        value_t* y, int n, value_t* y_unique, int n_classes, value_t* y_out, int idx, cudaStream_t stream) {
+    detail::getOvrLabels<value_t>(y, n, y_unique, n_classes, y_out, idx, stream);
 }
-
-// TODO: add one-versus-one selection: select two classes, relabel them to
-// +/-1, return array with the new class labels and corresponding indices.
-
-template <typename Type, int TPB_X, typename Lambda>
-__global__ void map_label_kernel(Type* map_ids,
-                                 size_t N_labels,
-                                 Type* in,
-                                 Type* out,
-                                 size_t N,
-                                 Lambda filter_op,
-                                 bool zero_based = false)
-{
-  int tid = threadIdx.x + blockIdx.x * TPB_X;
-  if (tid < N) {
-    if (!filter_op(in[tid])) {
-      for (size_t i = 0; i < N_labels; i++) {
-        if (in[tid] == map_ids[i]) {
-          out[tid] = i + !zero_based;
-          break;
-        }
-      }
-    }
-  }
-}
-
 /**
  * Maps an input array containing a series of numbers into a new array
  * where numbers have been mapped to a monotonically increasing set
@@ -157,18 +84,9 @@ __global__ void map_label_kernel(Type* map_ids,
  */
 template <typename Type, typename Lambda>
 void make_monotonic(
-  Type* out, Type* in, size_t N, cudaStream_t stream, Lambda filter_op, bool zero_based = false)
+        Type* out, Type* in, size_t N, cudaStream_t stream, Lambda filter_op, bool zero_based = false)
 {
-  static const size_t TPB_X = 256;
-
-  dim3 blocks(raft::ceildiv(N, TPB_X));
-  dim3 threads(TPB_X);
-
-  rmm::device_uvector<Type> map_ids(0, stream);
-  int num_clusters = getUniquelabels(map_ids, in, N, stream);
-
-  map_label_kernel<Type, TPB_X><<<blocks, threads, 0, stream>>>(
-    map_ids.data(), num_clusters, in, out, N, filter_op, zero_based);
+    detail::make_monotonic<Type, Lambda>(out, in, N, stream, filter_op, zero_based);
 }
 
 /**
@@ -191,8 +109,7 @@ void make_monotonic(
 template <typename Type>
 void make_monotonic(Type* out, Type* in, size_t N, cudaStream_t stream, bool zero_based = false)
 {
-  make_monotonic<Type>(
-    out, in, N, stream, [] __device__(Type val) { return false; }, zero_based);
+    detail::make_monotonic<Type>(out, in, N, stream, zero_based);
 }
 };  // namespace label
 };  // end namespace raft
