+++ cpp/include/raft/spectral/detail/spectral_util.hpp
