+++ cpp/include/raft/spectral/modularity_maximization.hpp
@@ -16,43 +16,12 @@
 
 #pragma once
 
-#include <math.h>
-#include <stdio.h>
-
-#include <cuda.h>
-#include <thrust/fill.h>
-#include <thrust/reduce.h>
-#include <thrust/transform.h>
-
 #include <tuple>
 
-#include <raft/spectral/cluster_solvers.hpp>
-#include <raft/spectral/eigen_solvers.hpp>
-#include <raft/spectral/spectral_util.hpp>
-
-#ifdef COLLECT_TIME_STATISTICS
-#include <cuda_profiler_api.h>
-#include <stddef.h>
-#include <sys/resource.h>
-#include <sys/sysinfo.h>
-#include <sys/time.h>
-#endif
-
-#ifdef COLLECT_TIME_STATISTICS
-static double timer(void)
-{
-  struct timeval tv;
-  cudaDeviceSynchronize();
-  gettimeofday(&tv, NULL);
-  return (double)tv.tv_sec + (double)tv.tv_usec / 1000000.0;
-}
-#endif
+#include <raft/spectral/detail/modularity_maximization.hpp>
 
 namespace raft {
 namespace spectral {
-
-using namespace matrix;
-using namespace linalg;
 
 // =========================================================
 // Spectral modularity_maximization
@@ -81,52 +50,15 @@ using namespace linalg;
  */
 template <typename vertex_t, typename weight_t, typename EigenSolver, typename ClusterSolver>
 std::tuple<vertex_t, weight_t, vertex_t> modularity_maximization(
-  handle_t const& handle,
-  sparse_matrix_t<vertex_t, weight_t> const& csr_m,
-  EigenSolver const& eigen_solver,
-  ClusterSolver const& cluster_solver,
-  vertex_t* __restrict__ clusters,
-  weight_t* eigVals,
-  weight_t* eigVecs)
-{
-  RAFT_EXPECTS(clusters != nullptr, "Null clusters buffer.");
-  RAFT_EXPECTS(eigVals != nullptr, "Null eigVals buffer.");
-  RAFT_EXPECTS(eigVecs != nullptr, "Null eigVecs buffer.");
-
-  auto stream   = handle.get_stream();
-  auto cublas_h = handle.get_cublas_handle();
-
-  std::tuple<vertex_t, weight_t, vertex_t>
-    stats;  // # iters eigen solver, cluster solver residual, # iters cluster solver
-
-  vertex_t n = csr_m.nrows_;
-
-  // Compute eigenvectors of Modularity Matrix
-
-  // Initialize Modularity Matrix
-  modularity_matrix_t<vertex_t, weight_t> B{handle, csr_m};
-
-  auto eigen_config = eigen_solver.get_config();
-  auto nEigVecs     = eigen_config.n_eigVecs;
-
-  // Compute eigenvectors corresponding to largest eigenvalues
-  std::get<0>(stats) = eigen_solver.solve_largest_eigenvectors(handle, B, eigVals, eigVecs);
-
-  // Whiten eigenvector matrix
-  transform_eigen_matrix(handle, n, nEigVecs, eigVecs);
-
-  // notice that at this point the matrix has already been transposed, so we are scaling
-  // columns
-  scale_obs(nEigVecs, n, eigVecs);
-  RAFT_CHECK_CUDA(stream);
-
-  // Find partition clustering
-  auto pair_cluster = cluster_solver.solve(handle, n, nEigVecs, eigVecs, clusters);
-
-  std::get<1>(stats) = pair_cluster.first;
-  std::get<2>(stats) = pair_cluster.second;
-
-  return stats;
+        handle_t const& handle,
+        sparse_matrix_t<vertex_t, weight_t> const& csr_m,
+        EigenSolver const& eigen_solver,
+        ClusterSolver const& cluster_solver,
+        vertex_t* __restrict__ clusters,
+        weight_t* eigVals,
+        weight_t* eigVecs) {
+    return detail::modularity_maximization<vertex_t, weight_t, EigenSolver, ClusterSolver>(handle, csr_m,
+                                                                                           eigen_solver, cluster_solver, clusters, eigVals, eigVecs);
 }
 //===================================================
 // Analysis of graph partition
@@ -144,42 +76,9 @@ void analyzeModularity(handle_t const& handle,
                        sparse_matrix_t<vertex_t, weight_t> const& csr_m,
                        vertex_t nClusters,
                        vertex_t const* __restrict__ clusters,
-                       weight_t& modularity)
-{
-  RAFT_EXPECTS(clusters != nullptr, "Null clusters buffer.");
-
-  vertex_t i;
-  vertex_t n = csr_m.nrows_;
-  weight_t partModularity, clustersize;
-
-  auto cublas_h = handle.get_cublas_handle();
-  auto stream   = handle.get_stream();
-
-  // Device memory
-  vector_t<weight_t> part_i(handle, n);
-  vector_t<weight_t> Bx(handle, n);
-
-  // Initialize cuBLAS
-  RAFT_CUBLAS_TRY(cublassetpointermode(cublas_h, CUBLAS_POINTER_MODE_HOST, stream));
-
-  // Initialize Modularity
-  modularity_matrix_t<vertex_t, weight_t> B{handle, csr_m};
-
-  // Initialize output
-  modularity = 0;
-
-  // Iterate through partitions
-  for (i = 0; i < nClusters; ++i) {
-    if (!construct_indicator(handle, i, n, clustersize, partModularity, clusters, part_i, Bx, B)) {
-      WARNING("empty partition");
-      continue;
-    }
-
-    // Record results
-    modularity += partModularity;
-  }
-
-  modularity = modularity / B.diagonal_.nrm1();
+                       weight_t& modularity) {
+    detail::analyzeModularity<vertex_t, weight_t>(handle, csr_m, nClusters, clusters, modularity);
+}
 }
 
 }  // namespace spectral
